Date now converts to overridden type
Also, fixed spelling of overriden -> overridden

An alternative to this would be just to bypass `$ref` whenever there is an overridden type.

It still throws the error. It happens at:
https://github.com/lbovet/typson/blob/master/lib/typson.js#L86

Not sure what solution should be used to silence the error. Could check if the value in `semanticDiagnostics` is in the array `overriddenTypes`. But I would need to create a new file that both files can call. Of course, we could just let the error pass through, just annoying though when a person is using error checking.

fixes #16